requirements.txt
streamlit>=1.28.0
pandas>=2.0.0
pyyaml>=6.0
altair>=5.0.0
pypdf>=3.0.0
PyPDF2>=3.0.0
